refactor: split out command factory
